Add files via upload
This is the source code for the Android app Samourai Wallet, version 0.99.98ii. This is only the source code and not the app itself, and this zip file is provided for educational purposes only, please don't do anything illegal with it. I own no copyrights, and do know that of the time of this posting 8/8/2024 the developers of the Samourai Wallet App are facing legal issues. Experiment with at your own risk.